Extract all the animation and param IDs from penguin animation component class into penguin component folder
