+++ Assets/Code/Game/Entities/Penguin/Components/PenguinAnimationEventId.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2620b6af104e5ee45a8fa438057ea843
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
