+++ Assets/Code/Game/Entities/Penguin/Components/PenguinAnimationParamId.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 45846b9e50b270243941a209db233129
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
